refactor(webview): extract profile picker module, shared ProfileSummary, core from-unknown utils; schema pins compat (#212)
