+++ .trellis/spec/frontend/hexscope-storage.md
@@ -71,7 +71,7 @@ Layout (whole tree git-tracked — no `.gitignore` seeding, no local/private spl
 └── scripts/                  # script runner, unchanged
 ```
 
-- **Three tiers**: struct *types* live once in the workspace pool (`.hexscope/structs.json`); a *profile* is a named annotation bundle (`ProfileRecord`) storing pins / activeChecks / endian / segmentNames / labels, not owned by any file; a *file* is bound to a profile by a `Binding` entry keyed on its workspace-relative posix `relPath`. The pool's in-memory fallback cache is **per root** (keyed by `root`, never a module global), so one root's defs can never leak as another root's empty default. All profiles live in the single `profiles.json` array (order preserved; ids unique — `normalizeProfilesRegistry` drops duplicates and malformed records). The old `firmware_profiles/*` per-file dirs (`index.json` + `structs.json` + `integrity.json`) and the per-dir registry (`profiles/<id>/profile.json`) no longer exist outside migration. The standalone `IntegrityProfile[]` template registry is removed — checks live only as `activeChecks` inside a `ProfileRecord` (the Integrity panel shows only the bound profile's `activeChecks`; see integrity-checks.md).
+- **Three tiers**: struct *types* live once in the workspace pool (`.hexscope/structs.json`); a *profile* is a named annotation bundle (`ProfileRecord`) storing structPins / activeChecks / endian / segmentNames / labels, not owned by any file; a *file* is bound to a profile by a `Binding` entry keyed on its workspace-relative posix `relPath`. The pool's in-memory fallback cache is **per root** (keyed by `root`, never a module global), so one root's defs can never leak as another root's empty default. All profiles live in the single `profiles.json` array (order preserved; ids unique — `normalizeProfilesRegistry` drops duplicates and malformed records). The old `firmware_profiles/*` per-file dirs (`index.json` + `structs.json` + `integrity.json`) and the per-dir registry (`profiles/<id>/profile.json`) no longer exist outside migration. The standalone `IntegrityProfile[]` template registry is removed — checks live only as `activeChecks` inside a `ProfileRecord` (the Integrity panel shows only the bound profile's `activeChecks`; see integrity-checks.md).
 - **Envelope**: every file is `{ version: 1, data }`. `writeJson`/`flush` write it; `readJson` unwraps. Unknown `version` → `corrupt` (empty default + warn once per store instance + file never overwritten — forward protection). Unversioned payloads (bare array/object) accepted as current version, lazily upgraded on next write. `$schema` lives at envelope level only, never in `data`.
 - **`$schema`**: on any write `writeJson` injects the matching `schemas/<name>.schema.json` sibling (keeps an existing string sibling if present). Normalizers never see it (`readJson` unwraps first) so self-heal cannot strip it. Top-level files reference `schemas/<name>.schema.json` directly (no `../../` — there are no profile subdirs anymore).
 - **Load semantics** (`JsonStore.load`): missing → `empty()`; corrupt/unknown-version → warn once + `empty()`; parse-ok → normalize; self-heal write-back only when `changed` (parse-ok AND normalized differs).
@@ -82,7 +82,7 @@ Layout (whole tree git-tracked — no `.gitignore` seeding, no local/private spl
 - **Watcher**: `attachProfileWatcher` watches `profiles.json`, `structs.json`, and `bindings.json` under `.hexscope/` (create/change/delete). Session debounces (`onProfileChanged` → per-slot `scheduleReload`) and guards self-writes via the 1 s self-write horizon. External edits auto-apply silently (see editing-save-external-change.md §1a).
 - **Webview fan-out** on external change: `profiles.json` → `perFileDataChange` (bound record changed) + `profilesState` (registry/binding changed); `.hexscope/structs.json` → `structsExternalChange` (webview prunes pins with vanished `structId`). Struct edits (`saveStructs`) always write the workspace pool; pin/label/check/endian edits always write the bound profile; picking/creating a profile writes a binding.
 - **Migration** (once per root, first panel open, before first `postInit`): (0) per-dir registry `.hexscope/profiles/<id>/profile.json` → merge into `profiles.json` array (dedupe ids + case-insensitive names; existing array records win), one-time Memento marker `hexScope.profilesArray.v1`, dir tree left in place for rollback (runs before the firmware_profiles conversion, which pushes more records into the array); (a) Memento-era `globalState`/`workspaceState` keys (structs v2/v1, integrity templates, per-file labels/segmentNames/pins/checks/endian/structs) → deduped pool + unbound template profiles + a bound profile for the open doc, then hard-delete keys; (b) on-disk `firmware_profiles/*` tree → merge every `structs.json` into the pool (dedup via structMigration `mergeLegacyStructDefs`) and convert each `index.json`+`integrity.json` pair into one registry profile + one binding (extra integrity templates → unbound profiles). Idempotent: Memento tariffs + per-converted-dir `.converted` marker file (restart-safe, no duplicate profiles); the mere existence of `bindings.json` is **never** treated as proof of completion, so a pre-existing binding from an unrelated file cannot suppress conversion of a remaining legacy tree. Legacy trees left in place for rollback. Failures never block opening.
-- **JSON Schemas**: repo `schemas/{profiles,structs,bindings}.schema.json` bundled; `contributes.jsonValidation` globs map the three file paths → bundled schema (`schemas/index.schema.json`, `schemas/integrity.schema.json`, `schemas/profile.schema.json` retired); seeded `.hexscope/schemas/` copies for installed-extension/agent resolution. `profiles.schema.json` data is an array (each item = the ProfileRecord shape; `uniqueItems` catches exact duplicates; id/name uniqueness beyond that is runtime-enforced by `normalizeProfilesRegistry`). Envelope root tolerant (`additionalProperties: true`), nested payload objects strict (`required` + `additionalProperties: false` + enums from `STRUCT_FIELD_TYPES` / `INTEGRITY_ALGORITHMS` / `endian`). Schemas describe the contracted shape; they do not gate loading (runtime stays lenient).
+- **JSON Schemas**: repo `schemas/{profiles,structs,bindings}.schema.json` bundled; `contributes.jsonValidation` globs map the three file paths → bundled schema (`schemas/index.schema.json`, `schemas/integrity.schema.json`, `schemas/profile.schema.json` retired); seeded `.hexscope/schemas/` copies for installed-extension/agent resolution. `profiles.schema.json` data is an array (each item = the ProfileRecord shape; `uniqueItems` catches exact duplicates; id/name uniqueness beyond that is runtime-enforced by `normalizeProfilesRegistry`). `structPins` stays `required`; the legacy pre-rename `pins` key is declared as an **allowed deprecated property** (same `structPin` array, `"deprecated": true`) so files still carrying it pass strict authoring while the runtime's `structPins ?? pins` tolerance keeps reading them. Envelope root tolerant (`additionalProperties: true`), nested payload objects strict (`required` + `additionalProperties: false` + enums from `STRUCT_FIELD_TYPES` / `INTEGRITY_ALGORITHMS` / `endian`). Schemas describe the contracted shape; they do not gate loading (runtime stays lenient).
 
 ### 4. Validation & Error Matrix
