+++ .trellis/spec/frontend/state-management.md
@@ -51,13 +51,15 @@ Reverse flow uses `WebviewToProviderMessage` through `postProviderMessage`. The 
 
 ## WebviewToProviderMessage additions (profile actions)
 
-`saveProfile` (explicit flush; materializes out-of-workspace), `duplicateProfile` (Save as…: copies the bound profile, binds the current file to the copy), `renameProfile`, `deleteProfile` (confirms with bound-file count). All run inside `enqueuePerFileOp` and end with `broadcastPerFileData()` + `broadcastProfilesState()`.
+`saveProfile` (explicit flush; materializes out-of-workspace), `duplicateProfile` (Save as…: copies the bound profile, binds the current file to the copy), `renameProfile`, `deleteProfile` (confirms with bound-file count). All run inside `enqueuePerFileOp` and end with `broadcastPerFileData()` + `broadcastProfilesState()`. In the webview these are the four `ProfileAction` commands of the typed `PROFILE_ACTION_MESSAGES: Record<ProfileAction, WebviewToProviderMessage>` map guarded by `isProfileAction(cmd)` in `src/webview/profilePicker.ts`.
+
+The webview profile payload shape is the shared `ProfileSummary { id; name }` (`src/webviewProtocol.ts`), used uniformly by the init/profile + `profilesState` message types, `S.profileState.profiles` (`state.ts`), `WebviewProfileState` in `webviewMessageModel.ts`, and the host's `listProfiles()` return type.
 
 ## On-disk JSON Schema contract
 
 - Three strict JSON Schemas describe the `.hexscope/` on-disk shapes so editors and AI agents can validate/author team state: `schemas/{profiles,structs,bindings}.schema.json` in the repo (bundled in the VSIX) and seeded into `.hexscope/schemas/` on the first registry write.
 - Every storage file is envelope `{ version: const 1, data, $schema? }`; `$schema` is a **relative path** to the workspace-seeded schema copy (`schemas/<name>.schema.json`) so terminal agents resolve the contract from the file itself. `readJson` unwraps the envelope before normalizing, so normalizers never see `$schema`; `writeJson` re-injects the canonical sibling on every write — self-heal cannot strip it.
-- Editors bind via `contributes.jsonValidation` globs (`.hexscope/profiles.json`, `.hexscope/structs.json`, `.hexscope/bindings.json`). Schemas are **strict for authoring** (`required` everywhere, enums for struct types / integrity algorithms / endian, nested `additionalProperties: false`, `profiles.json` data is an array with `uniqueItems`), but the runtime **remains lenient** — normalizers tolerate the extra fields the schema flags, corrupt/unknown-version files load empty + warn once, never overwrite. Id/name uniqueness beyond exact duplicates is runtime-enforced by `normalizeProfilesRegistry`.
+- Editors bind via `contributes.jsonValidation` globs (`.hexscope/profiles.json`, `.hexscope/structs.json`, `.hexscope/bindings.json`). Schemas are **strict for authoring** (`required` everywhere, enums for struct types / integrity algorithms / endian, nested `additionalProperties: false`, `profiles.json` data is an array with `uniqueItems`), but the runtime **remains lenient** — normalizers tolerate the extra fields the schema flags, corrupt/unknown-version files load empty + warn once, never overwrite. Id/name uniqueness beyond exact duplicates is runtime-enforced by `normalizeProfilesRegistry`. `profiles.schema.json` keeps `structPins` required and additionally declares the legacy `pins` key as a deprecated property so pre-rename files pass strict validation (runtime reads `structPins ?? pins`).
 - Drift guard: `ajv` (devDependency, test-only) validates fixtures in `src/test/schemas/schemaValidation.test.ts`, pinned to `DATA_VERSION` and the source enum consts.
 
 ## Update Pattern
